requirements.txt
python-telegram-bot==20.7
aiosqlite==0.19.0
aiohttp==3.9.1
APScheduler==3.10.4
python-dotenv==1.0.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.3
cachetools==5.3.2